refactor: PR 수정 요청 변경
