+++ src/main/java/com/fourweekdays/fourweekdays/inbound/exception/InboundException.java
@@ -1,4 +1,4 @@
-package com.fourweekdays.fourweekdays.Inbound.exception;
+package com.fourweekdays.fourweekdays.inbound.exception;
 
 import com.fourweekdays.fourweekdays.global.exception.BaseException;
 import com.fourweekdays.fourweekdays.global.exception.ExceptionType;
