+++ src/main/java/com/fourweekdays/fourweekdays/inbound/model/entity/InboundStatus.java
@@ -1,4 +1,4 @@
-package com.fourweekdays.fourweekdays.Inbound.model.entity;
+package com.fourweekdays.fourweekdays.inbound.model.entity;
 
 public enum InboundStatus {
     REQUESTED, RECEIVED, COMPLETED
