+++ src/main/java/com/fourweekdays/fourweekdays/inbound/repository/InboundRepository.java
@@ -1,6 +1,6 @@
-package com.fourweekdays.fourweekdays.Inbound.repository;
+package com.fourweekdays.fourweekdays.inbound.repository;
 
-import com.fourweekdays.fourweekdays.Inbound.model.entity.Inbound;
+import com.fourweekdays.fourweekdays.inbound.model.entity.Inbound;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InboundRepository extends JpaRepository<Inbound, Long> {
